Adapt extension spec tag tests
